+++ src/RecyclableCalculator.Core/Domain/Models/RecyclableType.cs
@@ -13,10 +13,10 @@ namespace RecyclableCalculator.Core.Domain.Models
 
 		public string Type { get; set; }
 
-		public double Rate { get; set; }
+		public decimal Rate { get; set; }
 
-		public double MinKg { get; set; }
+		public decimal MinKg { get; set; }
 
-		public double MaxKg { get; set; }
+		public decimal MaxKg { get; set; }
 	}
 }
